+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1
+
+* Forms can be re-enabled by updating the state `isSubmitting` on any form components.
+* Developers can now set the alignment on an input's value using the `align` prop.
+
 # 0.1.0
 
 ## New Components